Separate Blog Post component from template component
